// FCUBEAPP/src/app/models/docrenewalmastermodel.ts
export class Docrenewalmastermodel {
    docRenewalID: string = "";
    docCode: string = "";
    docDescription: string = "";
    reminderDays: string = "";
    debitType: string = "";
    debitAc: string = "";
    isActive: string = "";
    recurring_Onetime: string = "";
    loggedInUser: string = "";
}